feat: progress in LpNormRegularization
