+++ src/cofi/utils/__init__.py
@@ -1,17 +1,15 @@
 """Utility classes and functions (e.g. to generate regularization terms and more)
 """
 
-from ._regularization import (
-    BaseRegularization,
-    LpNormRegularization,
-    GaussianPrior,
-    QuadraticReg,
-)
+from ._reg_base import BaseRegularization
+from ._reg_lp_norm import LpNormRegularization, QuadraticReg
+from ._reg_model_cov import ModelCovariance, GaussianPrior
 
 
 __all__ = [
     "BaseRegularization",
     "LpNormRegularization",
-    "GaussianPrior",
     "QuadraticReg",
+    "ModelCovariance", 
+    "GaussianPrior",
 ]
