+++ src/cofi/utils/_reg_lp_norm.py
@@ -1,10 +1,10 @@
-from numbers import Number
-from abc import abstractmethod, ABCMeta
 from typing import Union
+from numbers import Number
 from functools import reduce
 import numpy as np
 import scipy
 
+from ._reg_base import BaseRegularization
 from .._exceptions import DimensionMismatchError
 
 
@@ -15,177 +15,26 @@ REG_TYPES = {
     "smoothing": 2,
 }
 
-
-class BaseRegularization(metaclass=ABCMeta):
-    r"""Base class for a regularization term
-
-    Check :class:`QuadraticReg` for a concrete example.
-
-    .. rubric:: Basic interface
-
-    The basic properties / methods for a regularization term in ``cofi.utils``
-    include the following:
-
-    .. autosummary::
-        BaseRegularization.model_size
-        BaseRegularization.reg
-        BaseRegularization.gradient
-        BaseRegularization.hessian
-        BaseRegularization.__call__
-
-    .. rubric:: Adding two terms
-
-    Two instances of ``BaseRegularization`` can also be added together using the ``+``
-    operator:
-
-    .. autosummary::
-        BaseRegularization.__add__
-
-    """
-
-    def __init__(
-        self,
-    ):
-        pass
-
-    @property
-    @abstractmethod
-    def model_shape(self) -> tuple:
-        """the shape of models that current regularization function accepts"""
-        raise NotImplementedError
-    
-    @property
-    def model_size(self) -> Number:
-        """the number of unknowns that current regularization function accepts"""
-        return reduce(lambda a, b: a * b, np.array(self.model_shape), 1)
-
-    def __call__(self, model: np.ndarray) -> Number:
-        r"""a class instance itself can also be called as a function
-
-        It works exactly the same as :meth:`reg`.
-
-        In other words, the following two usages are exactly the same::
-
-        >>> my_reg = QuadraticReg(factor=1, model_size=3)
-        >>> my_reg_value = my_reg(np.array([1,2,3]))            # usage 1
-        >>> my_reg_value = my_reg.reg(np.array([1,2,3]))        # usage 2
-        """
-        return self.reg(model)
-
-    @abstractmethod
-    def reg(self, model: np.ndarray) -> Number:
-        """the regularization function value given a model to evaluate"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def gradient(self, model: np.ndarray) -> np.ndarray:
-        """the gradient of regularization function with respect to model given a model
-
-        The usual size for the gradient is :math:`(M,)` where :math:`M` is the number
-        of model parameters
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def hessian(self, model: np.ndarray) -> np.ndarray:
-        """the hessian of regularization function with respect to model given a model
-
-        The usual size for the Hessian is :math:`(M,M)` where :math:`M` is the number
-        of model parameters
-        """
-        raise NotImplementedError
-
-    def __add__(self, other_reg):
-        r"""Adds two regularization terms
-
-        Parameters
-        ----------
-        other_reg : BaseRegularization
-            the second argument of "+" operator; must also be a
-            :class:`BaseRegularization` instance
-
-        Returns
-        -------
-        BaseRegularization
-            a regularization term ``resRegularization`` such that:
-
-            - :math:`\text{resRegularization.reg}(m)=\text{self.reg}(m)+\text{other_reg.reg}(m)`
-            - :math:`\text{resRegularization.gradient}(m)=\text{self.gradient}(m)+\text{other_reg.gradient}(m)`
-            - :math:`\text{resRegularization.hessian}(m)=\text{self.hessian}(m)+\text{other_reg.hessian}(m)`
-
-        Raises
-        ------
-        TypeError
-            when the ``other_reg`` is not a regularization term generated by CoFI Utils
-        DimensionMismatchError
-            when the ``other_reg`` doesn't accept model_size that matches the one of
-            ``self``
-
-        Examples
-        --------
-
-        >>> from cofi import BaseProblem
-        >>> from cofi.utils import QuadraticReg
-        >>> reg1 = QuadraticReg(factor=1, model_size=3, reg_type="damping")
-        >>> reg2 = QuadraticReg(factor=2, model_size=3, reg_type="smoothing")
-        >>> my_problem = BaseProblem()
-        >>> my_problem.set_regularization(reg1 + reg2)
-
-        """
-        if not isinstance(other_reg, BaseRegularization):
-            raise TypeError(
-                f"unsupported operand type(s) for +: '{self.__class__.__name__}' "
-                f"and '{other_reg.__class__.__name__}"
-            )
-        if self.model_size != other_reg.model_size:
-            raise DimensionMismatchError(
-                entered_name="the second regularization term",
-                entered_dimension=other_reg.model_size,
-                expected_source="the first regularization term",
-                expected_dimension=self.model_size,
-            )
-        tmp_model_shape = self.model_shape
-        tmp_reg = self.reg
-        tmp_grad = self.gradient
-        tmp_hess = self.hessian
-
-        class CompositeRegularization(BaseRegularization):
-            @property
-            def model_shape(self):
-                return tmp_model_shape
-
-            def reg(self, model):
-                return tmp_reg(model) + other_reg(model)
-
-            def gradient(self, model):
-                return tmp_grad(model) + other_reg.gradient(model)
-
-            def hessian(self, model):
-                return tmp_hess(model) + other_reg.hessian(model)
-
-        return CompositeRegularization()
-
-
 class LpNormRegularization(BaseRegularization):
     r"""CoFI's utility class to calculate Lp-norm regularization, given the p value
     (default to 2), an optional weighting matrix and an optional reference value
 
-    :math:`L(p, W, m_0) = ||W(m-m_0)||_p^p`
-    
+    :math:`L(p, W, m_0) = ||W(m-m_0)||_p^p = \sum_i |W(m-m_0)_i|^p`
+
     Parameters
     ----------
-    p : Number or str
-        order value passed directly to :func:`scipy.linalg.norm`, default to 2
+    p : Number
+        order value (p in the formula above), default to 2
     weighting_matrix: str or np.ndarray
-        regularization type (one of {:code:`"damping"`, :code:`"flattening"` 
+        regularization type (one of {:code:`"damping"`, :code:`"flattening"`
         :code:`"smoothing"`}), or a bring-your-own weighting matrix, default to
         "damping" (identity matrix for weighting)
     model_shape: tuple
-        shape of the model, must be supplied if the :code:`reference_model` is not 
+        shape of the model, must be supplied if the :code:`reference_model` is not
         given
     reference_model: np.ndarray
         :math:`m_0` in the formula above
-    
+
     Raises
     ------
     ValueError
@@ -194,14 +43,15 @@ class LpNormRegularization(BaseRegularization):
         if both :code:`model_size` and :code:`reference_model` are given but they don't
         match in dimension
     """
+
     def __init__(
-        self, 
-        p : Union[Number, str] = 2, 
-        weighting_matrix : Union[str, np.ndarray] ="damping", 
-        model_shape : tuple = None, 
-        reference_model : np.ndarray = None 
+        self,
+        p: Number = 2,
+        weighting_matrix: Union[str, np.ndarray] = "damping",
+        model_shape: tuple = None,
+        reference_model: np.ndarray = None,
     ):
-        self._order = p
+        self._order = self._validate_p(p)
         self._weighting_matrix = weighting_matrix
         self._model_shape = model_shape
         self._reference_model = reference_model
@@ -238,10 +88,10 @@ class LpNormRegularization(BaseRegularization):
 
         if isinstance(self._weighting_matrix, str) or self._weighting_matrix is None:
             _reg_type = self._weighting_matrix
-            if _reg_type == "damping" or _reg_type is None:          # 0th order difference
+            if _reg_type == "damping" or _reg_type is None:  # 0th order difference
                 self._weighting_matrix = np.identity(self.model_size)
-            elif _reg_type in REG_TYPES:         # 1st/2nd order difference
-                if np.size(self.model_shape) == 1:              # 1D model
+            elif _reg_type in REG_TYPES:  # 1st/2nd order difference
+                if np.size(self.model_shape) == 1:  # 1D model
                     order = REG_TYPES[_reg_type]
                     if self.model_size < order + 2:
                         raise ValueError(
@@ -250,7 +100,9 @@ class LpNormRegularization(BaseRegularization):
                         )
                     d_dx = findiff.FinDiff(0, 1, order)
                     self._weighting_matrix = d_dx.matrix((self.model_size,))
-                elif np.size(self.model_shape) == 2 and np.ndim(self.model_shape) == 1:     # 2D model
+                elif (
+                    np.size(self.model_shape) == 2 and np.ndim(self.model_shape) == 1
+                ):  # 2D model
                     nx = self.model_shape[0]
                     ny = self.model_shape[1]
                     order = REG_TYPES[_reg_type]
@@ -263,12 +115,14 @@ class LpNormRegularization(BaseRegularization):
                     d_dy = findiff.FinDiff(1, 1, order)  # y direction
                     matx = d_dx.matrix((nx, ny))  # scipy sparse matrix
                     maty = d_dy.matrix((nx, ny))  # scipy sparse matrix
-                    self._weighting_matrix = np.vstack((matx.toarray(), maty.toarray()))  # combine above
+                    self._weighting_matrix = np.vstack(
+                        (matx.toarray(), maty.toarray())
+                    )  # combine above
                 else:
                     raise NotImplementedError(
                         "only 1D and 2D derivative operators implemented"
                     )
-        elif self._weighting_matrix:            # byo matrix
+        elif self._weighting_matrix:  # byo matrix
             if len(self._weighting_matrix.shape) != 2:
                 raise ValueError(
                     "the bring-your-own regularization matrix must be 2-dimensional"
@@ -283,6 +137,14 @@ class LpNormRegularization(BaseRegularization):
                 "please specify the weighting matrix either via a string among "
                 "\{`damping`, `flattening`, `smoothing`\}, or bringing your own matrix"
             )
+
+    @staticmethod
+    def _validate_p(p):
+        if not isinstance(p, Number):
+            raise ValueError(
+                f"number expected for argument p but got {p} of type {type(p)}"
+            )
+        return p
 
     def _validate_model(self, model):
         flat_m = np.ravel(model)
@@ -303,36 +165,6 @@ class LpNormRegularization(BaseRegularization):
 
     def _lp_norm(self, mat):
         return np.sum(np.abs(mat) ** self._order)
-
-
-class GaussianPrior(BaseRegularization):
-    r"""CoFI's utility class to calculate the Gaussian prior, given the inverse of
-    model covariance matrix and the mean model
-
-    :math:`GaussianPrior(C_m^{-1}, \mu) = -\frac{1}{2}(m-\mu)^TC_m^{-1}(m-\mu)`
-
-    Parameters
-    ----------
-    TODO
-    """
-
-    def __init__(self, model_covariance_inv, mean_model):
-        self._Cminv = model_covariance_inv
-        self._mu = mean_model
-        self._model_shape = mean_model.shape
-
-    def reg(self, model: np.ndarray) -> Number:
-        pass
-
-    def gradient(self, model: np.ndarray) -> np.ndarray:
-        return super().gradient(model)
-
-    def hessian(self, model: np.ndarray) -> np.ndarray:
-        return super().hessian(model)
-
-    @property
-    def model_shape(self) -> tuple:
-        return self._model_shape
 
 
 class QuadraticReg(BaseRegularization):
